Remove old frontend and refactor circuit in
