+++ src/Pipeline/Frontend/Circuit.hs
@@ -6,91 +6,11 @@
 module Pipeline.Frontend.Circuit where
 
 import Pipeline.Core.DataStore
--- import Pipeline.Core.Task
-import Data.Typeable (Typeable)
-import Prelude hiding (id, replicate)
-
-{-|
-  The main wrapping data type for a function. This makes working with the function type easier. 
--}
-data Task fs as g b = (
-  DataSource' fs as (Apply fs as),
-  DataSource g b,
-  Typeable fs, Typeable g,
-  Typeable as, Typeable b)
-  => Task (HList (Apply fs as) -> g b -> IO (g b)) (g b)
+import Pipeline.Core.Task
+import Prelude hiding (id, replicate, (<>))
 
 
--- |Required to store tasks of differing types in a single 'Map'. Uses existential types.
--- data TaskWrap = forall f a g b. (
---   DataSource f a, DataSource g b,
---   Typeable f, Typeable a, Typeable g, Typeable b) => TaskWrap (Task f a g b)
-
-{-|
-  This allows a function to be converted into a Task. 
--}
-multiInputFunctionTask :: (DataSource' fs as (Apply fs as), DataSource g b, Typeable as, Typeable b, Typeable fs, Typeable g) => (HList as -> b) -> g b -> Task fs as g b 
-multiInputFunctionTask f = Task (\sources sink -> do
-  input <- (hSequence . fetch') sources
-  save sink (f input))
-
-functionTask :: (DataSource f a, DataSource g b, Typeable f, Typeable a, Typeable g, Typeable b) => (a -> b) -> g b -> Task '[f] '[a] g b
--- It is okay to pattern match the hlist to just one value, as the type states that it only consumes one element.
-functionTask f = multiInputFunctionTask (\(HCons inp HNil) -> f inp)
-
-hSequence :: IOList as -> IO (HList as)
-hSequence IONil = return HNil
-hSequence (IOCons x xs) = do
-  x' <- x
-  xs' <- hSequence xs
-  return $ x' `HCons` xs'
-
-
-
-data IOList (xs :: [*]) where
-  IOCons :: IO x -> IOList xs -> IOList (x ': xs)
-  IONil :: IOList '[]
-
-data HList (xs :: [*]) where
-  HCons :: x -> HList xs -> HList (x ': xs)
-  HNil :: HList '[]
-  
-
-type family HAppendListR (l1 :: [k]) (l2 :: [k]) where
-  HAppendListR '[] l = l
-  HAppendListR (e ': l) l' = e ': HAppendListR l l'
-
-
-type family Apply (fs :: [* -> *]) (as :: [*]) = fas | fas -> fs as where
-  Apply '[] '[] = '[]
-  Apply (f ': fs) (a ': as) = f a ': Apply fs as
-
-
-class (xs ~ Apply fs as) => DataSource' (fs :: [* -> *]) (as :: [*]) (xs :: [*]) where
-  -- | Fetch the value stored in the 'DataSource'
-  -- fetch :: f a -> IO a
-  fetch' :: HList xs -> IOList as
-  -- | Save a value into the 'DataStore'
-  --   First argument depends on the instance. It may be 'empty' or it could be a pointer to a storage location.
-  -- save :: f a -> a -> IO (f a)
-  save' :: HList xs -> HList as -> IOList xs
-
--- -- for the user to define.
--- class DataSource f a where
---   fetch :: f a -> IO a
---   save :: f a -> a -> IO (f a)
-
-
-instance {-# OVERLAPPING #-} (x ~ f a, DataSource f a) => DataSource' '[f] '[a] '[x] where
-  fetch' (HCons x HNil) = IOCons (fetch x) IONil
-  save' = undefined
-
-instance (x ~ f a, DataSource f a, DataSource' fs as xs) => DataSource' (f ': fs) (a ': as) (x ': xs) where
-  fetch' (HCons x xs) = IOCons (fetch x) (fetch' xs)
-  save' = undefined
-
-
-
+-- Main type of the DSL
 data Circuit i o where
   Id        :: (DataSource' '[f] '[a] '[f a]) => Circuit (Apply '[f] '[a]) (Apply '[f] '[a])
   Apply     :: (DataSource' fs as (Apply fs as), DataSource' '[g] '[b] '[g b] ) => Task fs as g b -> Circuit (Apply fs as) (Apply '[g] '[b])
@@ -118,7 +38,7 @@ replicate :: (DataSource' '[f] '[a] '[f a]) => Circuit (Apply '[f] '[a]) (Apply 
              => Circuit (Apply fs as) (Apply gs bs)
              -> Circuit (Apply gs bs) (Apply hs cs)
              -> Circuit (Apply fs as) (Apply hs cs)
-(<|>)     :: (DataSource' fs as (Apply fs as), DataSource' gs bs (Apply gs bs), DataSource' hs cs (Apply hs cs), DataSource' is ds (Apply is ds))
+(<>)      :: (DataSource' fs as (Apply fs as), DataSource' gs bs (Apply gs bs), DataSource' hs cs (Apply hs cs), DataSource' is ds (Apply is ds))
              => Circuit (Apply fs as) (Apply gs bs)
              -> Circuit (Apply hs cs) (Apply is ds)
              -> Circuit (Apply (HAppendListR fs hs) (HAppendListR as cs)) (Apply (HAppendListR gs is) (HAppendListR bs ds))
@@ -129,19 +49,18 @@ id        = Id
 apply     = Apply
 replicate = Replicate
 (<->)     = Then
-(<|>)     = Beside
+(<>)      = Beside
 swap      = Swap
 dropL     = DropL
 dropR     = DropR
 
 infixr 4 <->
-infixr 5 <|>
+infixr 5 <>
   
 
 
--- This stuff is what I want
 
-
+-- Some example tasks
 plus1Task :: Task '[VariableStore] '[Int] VariableStore Int
 plus1Task = functionTask (+1) Empty
 
@@ -152,10 +71,13 @@ appendTask :: Task '[VariableStore, VariableStore] '[String, String] VariableSto
 appendTask = multiInputFunctionTask (\(HCons x (HCons y HNil)) -> x ++ y ) Empty
 
 
+-- Example Circuits
+
 -- Example1
 --   a  
 --   |  
---  +1  
+--  +1
+--   |
 --  show
 --   |  
 --   b
@@ -181,7 +103,7 @@ example2 :: Circuit '[VariableStore Int] '[VariableStore String, VariableStore I
 -- example2 = Then (Then Replicate (Beside (Apply plus1Task) (Apply showTask))) Swap
 example2 = replicate
            <->
-           apply plus1Task <|> apply showTask
+           apply plus1Task <> apply showTask
            <->
            swap
 
@@ -203,86 +125,11 @@ example2 = replicate
 --    |      |
 --    c      d
 example3 :: Circuit '[VariableStore Int, VariableStore Int] '[VariableStore String, VariableStore Int]
-example3 = example1         <|> example2
+example3 = example1         <> example2
            -- VariableStore String (1), VariableStore String (2), VariableStore Int
            <-> 
-           swap             <|> id
+           swap             <> id
            -- VariableStore String (2), VariableStore String (1), VariableStore Int
            <->
-           apply appendTask <|> id
+           apply appendTask <> id
            -- VariableStore String (2 ++ 1), VariableStore Int
-
-
-           
-
-
-
-
-
--- id :: DataSource f a => Circuit (f a) (f a)
--- id = ID
-
--- apply :: (DataSource f a, DataSource g b) => Task f a g b -> Circuit (f a) (g b)
--- apply = Apply
-
--- branch :: (DataSource f a) => Circuit (f a) (f a, f a)
--- branch = Branch
-
--- (<..>) :: (DataSource f a, DataSource g b, DataSource h c) => Circuit (f a) (g b) -> Circuit (g b) (h c) -> Circuit (f a) (h c)
--- (<..>) = Then
-
--- infixr 4 <..> 
-
-
--- (<&>) :: (DataSource f a, DataSource g b, DataSource h c, DataSource i d) => Circuit (f a) (g b) -> Circuit (h c) (i d) -> Circuit (f a, h c) (g b, i d)
--- (<&>) = Beside
-
--- infixr 5 <&>
-
--- swap :: (DataSource f a, DataSource g b) => Circuit (f a, g b) (g b, f a)
--- swap = Swap
-
-
-
-
-
-
--- example :: Circuit (VariableStore Int) (VariableStore String, VariableStore Int)
--- example = branch
---           <..>
---           apply plus1Task <&> id
---           <..>
---           apply plus1Task <&> apply showTask
---           <..>
---           swap
-
--- data VariableStore a = Var a | VEmpty
-
--- instance DataSource VariableStore a where
---   fetch (Var x) = return x
---   fetch VEmpty   = error "empty source"
---   save _ x = return (Var x)
-
--- instance DataSource (f :&&: g) a b where
---   fetch (x :&&: y) = do
---     x' <- fetch x 
---     y' <- fetch y
---     return (x', y')
---   save (x :&&: y) (x', y') = do
---     x'' <- save x x'
---     y'' <- save y y'
---     return x'' :&&: y''
-  
-
--- example' :: Circuit (VariableStore Int) (VariableStore (String, Int))
--- example' = branch
---            <..>
---            (apply (+1) <..> apply (+1)) <&> apply (show :: Int -> String)
---            <..>
---            swap
-
-  
--- example = Then (Then
---                  Branch
---                  (Apply (+1) `Beside` ID))
---                (Apply (+1) `Beside` Apply (+1))
