--- src/Pipeline/Frontend/PID.hs
@@ -1,5 +1,0 @@
-module Pipeline.Frontend.PID where
-
-
--- |A unique identifier for each process in a pipeline
-type PID = Int
